moved robot
